+++ app/widgets/VisioLink/VisioLink.php
@@ -10,10 +10,10 @@ class VisioLink extends Base
         $this->addcss('visiolink.css');
     }
 
-    public function ajaxDecline($to)
+    public function ajaxDecline($to, $id)
     {
         $this->rpc('Notification.incomingAnswer');
         $visio = new Visio;
-        $visio->ajaxTerminate($to, 'decline');
+        $visio->ajaxTerminate($to, 'decline', $id);
     }
 }